+++ test/cljs/cxengage_javascript_sdk/modules/reporting_test.cljs
@@ -6,6 +6,9 @@
             [cxengage-javascript-sdk.domain.rest-requests :as rest]
             [cljs-sdk-utils.interop-helpers :as ih]
             [cljs-sdk-utils.api :as api]
+            [cljs-sdk-utils.test :refer [camels]]
+            [cljs-sdk-utils.topics :as topics]
+            [cljs-sdk-utils.errors :as e]
             [cxengage-javascript-sdk.state :as st]
             [cxengage-javascript-sdk.modules.reporting :as rep]
             [cljs-uuid-utils.core :as uuid]
@@ -22,22 +25,42 @@
    :api-response {:results {:stat-one {:status 200
                                        :body {:results {:count 1}}}}}})
 
+(def not-found {:status 404})
+
 (deftest stat-query--happy-test--query-response-pubsub
   (testing "stat query function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-stat-query-response [:api-response :results])]
-                 (a/>! resp-chan successful-stat-query-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [_] resp-chan))
-                 (p/subscribe "cxengage/reporting/get-stat-query-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/stat-query {:statistic "resources-logged-in-count"}))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/batch-request (fn [& _]
+                                (go successful-stat-query-response)))
+     (p/subscribe
+      (topics/get-topic :get-stat-query-response)
+      (fn [error topic response]
+        (is (= (get-in successful-stat-query-response [:api-response :results]) (js->clj response :keywordize-keys true)))
+        (done)))
+     (rep/stat-query {:statistic "resources-logged-in-count"}))))
+
+(def resource-id (str (uuid/make-random-uuid)))
+(def queue-id (str (uuid/make-random-uuid)))
+(def mock-stat-bundle {:statistic "average-unit-test-time"
+                       :resource-id resource-id
+                       :queue-id queue-id})
+
+(deftest stat-query-error-test
+  (testing "stat query function success"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (reset! st/sdk-state {})
+     (set! rest/batch-request (fn [& _]
+                                (go not-found)))
+     (p/subscribe
+      (topics/get-topic :get-stat-query-response)
+      (fn [error topic response]
+        (is (= (camels (e/failed-to-perform-stat-query-err mock-stat-bundle not-found)) (js->clj error :keywordize-keys true)))
+        (done)))
+     (rep/stat-query mock-stat-bundle))))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; Get Capacity Tests
@@ -47,46 +70,63 @@
   {:status 200
    :api-response {:results {:resource-capacity []}}})
 
-(def tenant-capacity-url (str (st/get-base-api-url) "tenants/f5b660ef-9d64-47c9-9905-2f27a74bc14c/realtime-statistics/resource-capacity"))
-(def resource-capacity-url (str (st/get-base-api-url) "tenants/f5b660ef-9d64-47c9-9905-2f27a74bc14c/users/3e5890f1-0fef-46e3-b59f-3271e3d83646/realtime-statistics/resource-capacity"))
-
 (deftest get-capacity--happy-test--tenant
   (testing "get tenant capacity function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-capacity-response [:api-response :results])]
-                 (a/>! resp-chan successful-capacity-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [request]
-                                          (is (= (get request :url) tenant-capacity-url))
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/get-capacity-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/get-capacity))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-capacity-request (fn [& _]
+                                       (go successful-capacity-response)))
+     (p/subscribe
+      (topics/get-topic :get-capacity-response)
+      (fn [error topic response]
+        (is (= {:resourceCapacity []} (js->clj response :keywordize-keys true)))
+        (done)))
+     (rep/get-capacity))))
+
+(deftest get-capacity-error-response-tenant
+  (testing "get tenant capacity function error"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-capacity-request (fn [& _]
+                                       (go not-found)))
+     (p/subscribe
+      (topics/get-topic :get-capacity-response)
+      (fn [error topic response]
+        (is (= (camels (e/failed-to-get-capacity-err not-found)) (js->clj error :keywordize-keys true)))
+        (done)))
+     (rep/get-capacity))))
 
 (deftest get-capacity--happy-test--resource
   (testing "get resource capacity function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-capacity-response [:api-response :results])]
-                 (a/>! resp-chan successful-capacity-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [request]
-                                          (is (= (get request :url) resource-capacity-url))
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/get-capacity-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/get-capacity {:resource-id "3e5890f1-0fef-46e3-b59f-3271e3d83646"}))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-capacity-request (fn [& _]
+                                       (go successful-capacity-response)))
+     (p/subscribe
+      (topics/get-topic :get-capacity-response)
+      (fn [error topic response]
+        (is (= {:resourceCapacity []} (js->clj response :keywordize-keys true)))
+        (done)))
+     (rep/get-capacity {:resource-id resource-id}))))
+
+(deftest get-capacity-error-response-resource
+  (testing "get tenant capacity function error"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (let [old-request rest/get-capacity-request]
+       (set! rest/get-capacity-request (fn [& _]
+                                         (go not-found)))
+       (p/subscribe
+        (topics/get-topic :get-capacity-response)
+        (fn [error topic response]
+          (is (= (camels (e/failed-to-get-capacity-err not-found)) (js->clj error :keywordize-keys true)))
+          (set! rest/get-capacity-request old-request)
+          (done)))
+       (rep/get-capacity {:resource-id "3e5890f1-0fef-46e3-b59f-3271e3d83646"})))))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; Get Available Stats Tests
@@ -100,21 +140,30 @@
 
 (deftest get-available-stats--happy-test
   (testing "get available stats function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-available-response [:api-response])]
-                 (a/>! resp-chan successful-available-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [_]
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/get-available-stats-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/get-available-stats))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-available-stats-request (fn []
+                                              (go successful-available-response)))
+     (p/subscribe
+      (topics/get-topic :get-available-stats-response)
+      (fn [error topic response]
+        (is (= (camels (:api-response successful-available-response)) (js->clj response :keywordize-keys true)))
+        (done)))
+     (rep/get-available-stats))))
+
+(deftest get-available-stats-error-test
+  (testing "get-available-stats"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-available-stats-request (fn [] (go not-found)))
+     (p/subscribe
+      (topics/get-topic :get-available-stats-response)
+      (fn [error topic response]
+        (is (= (camels (e/failed-to-get-available-stats-err not-found)) (js->clj error :keywordize-keys true)))
+        (done)))
+     (rep/get-available-stats))))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; Get Interaction Tests
@@ -126,34 +175,41 @@
                             :channel-type "voice"
                             :customer "+1234567890"}}})
 
+(def interaction-id (str (uuid/make-random-squuid)))
+
 (deftest get-interaction--happy-test
   (testing "get interaction function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-interaction-response [:api-response])]
-                 (a/>! resp-chan successful-interaction-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [_]
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/get-interaction-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/get-interaction {:interaction-id "2937ac8b-380d-472b-9b9e-599097ee8c0d"}))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-interaction-history-request (fn [& _]
+                                                  (go successful-interaction-response)))
+     (p/subscribe
+      (topics/get-topic :get-interaction-response)
+      (fn [error topic response]
+        (is (= (camels (:api-response successful-interaction-response)) (js->clj response :keywordize-keys true)))
+        (done)))
+     (rep/get-interaction {:interaction-id interaction-id}))))
+
+(deftest get-interaction-error-test
+  (testing "get interaction function error"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-interaction-history-request (fn [& _]
+                                                  (go not-found)))
+     (p/subscribe
+      (topics/get-topic :get-interaction-response)
+      (fn [error topic response]
+        (is (= (camels (e/failed-to-get-interaction-reporting-err interaction-id not-found)) (js->clj error :keywordize-keys true)))
+        (done)))
+     (rep/get-interaction {:interaction-id interaction-id}))))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; Get Contact Interaction History Tests
 ;; -------------------------------------------------------------------------- ;;
 
-(def contact-history-url
-  (str (st/get-base-api-url)
-       "tenants/f5b660ef-9d64-47c9-9905-2f27a74bc14c/contacts/7749c9c0-3979-11e7-b8fc-d0f69d796523/interactions"))
-(def paged-history-url
-  (str (st/get-base-api-url)
-       "tenants/f5b660ef-9d64-47c9-9905-2f27a74bc14c/contacts/7749c9c0-3979-11e7-b8fc-d0f69d796523/interactions?page=5"))
+(def contact-id (str (uuid/make-random-squuid)))
 
 (def successful-contact-history-response
   {:status 200
@@ -161,43 +217,31 @@
 
 (deftest get-contact-history--happy-test
   (testing "get contact interaction history function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-contact-history-response
-                                                      [:api-response])]
-                 (a/>! resp-chan successful-contact-history-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [request]
-                                          (is (= (get request :url) contact-history-url))
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/get-contact-interaction-history-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/get-contact-interaction-history
-                  {:contact-id "7749c9c0-3979-11e7-b8fc-d0f69d796523"}))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-contact-interaction-history-request (fn [& _]
+                                                          (go successful-contact-history-response)))
+     (p/subscribe
+      (topics/get-topic :get-contact-interaction-history-response)
+      (fn [error topic response]
+        (is (= (:api-response successful-contact-history-response) (js->clj response :keywordize-keys true)))
+        (done)))
+     (rep/get-contact-interaction-history {:contact-id contact-id}))))
 
-(deftest get-contact-history--happy-test--paged
-  (testing "get paged contact interaction history function success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-contact-history-response [:api-response])]
-                 (a/>! resp-chan successful-contact-history-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [request]
-                                          (is (= (get request :url) paged-history-url))
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/get-contact-interaction-history-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (done)))
-                 (rep/get-contact-interaction-history {:contact-id "7749c9c0-3979-11e7-b8fc-d0f69d796523" :page 5}))))))
+(deftest get-contact-interaction-history-error-test
+  (testing "get-contact-interaction-history error response"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/get-contact-interaction-history-request (fn [& _]
+                                                          (go not-found)))
+     (p/subscribe
+      (topics/get-topic :get-contact-interaction-history-response)
+      (fn [error topic response]
+        (is (= (camels (e/failed-to-get-contact-interaction-history-err contact-id not-found)) (js->clj error :keywordize-keys true)))
+        (done)))
+     (rep/get-contact-interaction-history {:contact-id contact-id :page 5}))))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; Add Statistic Subscription Tests
@@ -205,81 +249,83 @@
 
 (def successful-batch-response
   {:status 200
-   :api-response {:results { :queue-length {:name "queue-length"
-                                            :type "interaction-count"
-                                            :user-friendly-name "Queue Length"}}}})
+   :api-response {:results {:queue-length {:name "queue-length"
+                                           :type "interaction-count"
+                                           :user-friendly-name "Queue Length"}}}})
 
 (def successful-stat-subs-update
   {:statistics {"c82d912c-2034-4b9e-a92a-f175870f5d8b" {:statistic "queue-length"
                                                         :topic "cxengage/reporting/stat-subscription-added"}}})
 
-(def successful-stat-sub-response
-  {:stat-id "c82d912c-2034-4b9e-a92a-f175870f5d8b"})
-
-(def stat-subs (atom {}))
-
 (deftest add-stat-sub--happy-test--batch-response
   (testing "add statistic subscription - batch success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)
-                     pubsub-expected-response (get-in successful-batch-response [:api-response :results])]
-                 (a/>! resp-chan successful-batch-response)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [_]
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/batch-response"
-                              (fn [error topic response]
-                                (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! api/api-request old)
-                                (reset! stat-subs)
-                                (done)))
-                 (rep/add-stat-subscription {:statistic "queue-length"}))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (set! rest/batch-request  (fn [& _]
+                                 (go successful-batch-response)))
+     (p/subscribe
+      (topics/get-topic :batch-response)
+      (fn [error topic response]
+        (is (= (get-in successful-batch-response [:api-response :results]) (ih/kebabify response)))
+        (done)))
+     (rep/add-stat-subscription {:statistic "queue-length"}))))
+
+(deftest add-stat-sub-error-test
+  (testing "Add statistic subscription, batch failure"
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (reset! rep/stat-subscriptions {})
+     (let [old-id uuid/make-random-uuid]
+       (set! uuid/make-random-uuid (fn [] "c82d912c-2034-4b9e-a92a-f175870f5d8b"))
+       (set! rest/batch-request (fn [& _]
+                                  (go not-found)))
+       (p/subscribe
+        (topics/get-topic :batch-response)
+        (fn [error topic response]
+          (is (= (camels (e/reporting-batch-request-failed-err (:statistics successful-stat-subs-update) not-found)) (js->clj error :keywordize-keys true)))
+          (set! uuid/make-random-uuid old-id)
+          (done)))
+       (rep/add-stat-subscription {:statistic "queue-length"})))))
 
 (deftest add-stat-sub--happy-test--subscription-added
   (testing "add statistic subscription - subscription success"
-    (async done
-           (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)]
-                 (a/>! resp-chan successful-stat-sub-response)
-                 (set! rep/stat-subscriptions stat-subs)
-                 (set! uuid/make-random-uuid #(str "c82d912c-2034-4b9e-a92a-f175870f5d8b"))
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [_]
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/stat-subscription-added"
-                              (fn [error topic response]
-                                (is (= successful-stat-sub-response (ih/kebabify response)))
-                                (is (= @stat-subs successful-stat-subs-update))
-                                (set! api/api-request old)
-                                (reset! stat-subs)
-                                (done)))
-                 (rep/add-stat-subscription {:statistic "queue-length"}))))))
+    (async
+     done
+     (reset! p/sdk-subscriptions {})
+     (let [old-id uuid/make-random-uuid]
+       (set! uuid/make-random-uuid (fn [] "c82d912c-2034-4b9e-a92a-f175870f5d8b"))
+       (p/subscribe
+        (topics/get-topic :add-stat)
+        (fn [error topic response]
+          (is (= {:statId "c82d912c-2034-4b9e-a92a-f175870f5d8b"}  (js->clj response :keywordize-keys true)))
+          (set! uuid/make-random-uuid old-id)
+          (done))))
+     (rep/add-stat-subscription {:statistic "queue-length"}))))
 
 ;; -------------------------------------------------------------------------- ;;
 ;; Remove Statistic Subscription Tests
 ;; -------------------------------------------------------------------------- ;;
 
-(def successful-stat-removal
-  {:statistics nil})
+(def stat-removal-mock-subscription
+  {:statistics {"c82d912c-2034-4b9e-a92a-f175870f5d8b" {:statistic "queue-length"
+                                                        :topic "cxengage/reporting/stat-subscription-added"}
+                "7749c9c0-3979-11e7-b8fc-d0f69d796523" {:statistic "queue-time"
+                                                        :topic "cxengage/reporting/stat-subscription-added"}}})
+
+(def stat-removal-response {:7749c9c0-3979-11e7-b8fc-d0f69d796523 {:statistic "queue-time"
+                                                                   :topic "cxengage/reporting/stat-subscription-added"}})
 
 (deftest remove-stat-sub--happy-test
   (testing "add statistic subscription - subscription success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old api/api-request
-                     resp-chan (a/promise-chan)]
-                 (a/>! resp-chan successful-stat-sub-response)
-                 (set! rep/stat-subscriptions stat-subs)
-                 (reset! st/sdk-state test-state)
-                 (set! api/api-request (fn [_]
-                                          resp-chan))
-                 (p/subscribe "cxengage/reporting/stat-subscription-removed"
-                              (fn [error topic response]
-                                (is (= @stat-subs successful-stat-removal))
-                                (set! api/api-request old)
-                                (reset! stat-subs)
-                                (done)))
-                 (rep/remove-stat-subscription {:stat-id "c82d912c-2034-4b9e-a92a-f175870f5d8b"}))))))
+           (reset! st/sdk-state test-state)
+           (reset! rep/stat-subscriptions stat-removal-mock-subscription)
+           (p/subscribe
+            (topics/get-topic :remove-stat)
+            (fn [error topic response]
+              (is (= stat-removal-response (js->clj response :keywordize-keys true)))
+              (done)))
+           (rep/remove-stat-subscription {:stat-id "c82d912c-2034-4b9e-a92a-f175870f5d8b"}))))